Flask_AI_app/requirements.txt
flask
numpy
pillow
tensorflow==2.13.0
gunicorn